==> notebooks/P1_identify_aggregate_properties.ipynb <==
# Establish connections to the input db and variables spreadsheets

# Import modules
import os, re
import arcpy
import pandas as pd
import numpy as np
from arcgis.features import GeoAccessor, GeoSeriesAccessor
from dotenv import load_dotenv

# Lot the .env variables
load_dotenv()

# Set the data paths
inputs_path = os.getenv('J111_INPUTS')
spatial_path = os.getenv('J111_SPATIAL')
env_path = os.getenv('J111_ENV')

# Set path to the input variables excel workbook
vars_xlsx = os.path.join(inputs_path, 'Redland_GAM_Input_Vars.xlsx')

# Get full path of Inputs GDB
input_gdb_path = os.path.join(spatial_path, 'redland_gam_current_dev.gdb')

# Set arcpy workspace
arcpy.env.workspace = os.path.join(env_path, 'J111_redland_gam.gdb')

# Import the input DCDB Base features into a dataframe
parcels = pd.DataFrame.spatial.from_featureclass(os.path.join(input_gdb_path, 'INPUT_FC_DCDB_Base')).drop(['OBJECTID', 'SHAPE'], axis=1)

# Import the Council Property features into dataframe
land = pd.DataFrame.spatial.from_featureclass(os.path.join(input_gdb_path, 'INPUT_FC_CouncilProperty')).drop(['OBJECTID', 'SHAPE'], axis=1)

# Inspect dataframe
parcels.head()

''' 
----------------------------------
Prepare the parcels to council land relational dataframe
----------------------------------
'''
# Import the Council Property to DCDB Base lots relational list
parcels_to_land = pd.DataFrame.spatial.from_table(os.path.join(input_gdb_path, 'INPUT_TB_DCDB_CouncilPropertyRel')).drop(['OBJECTID'], axis=1)

# Select only the required column
parcels_land = parcels_to_land.loc[:, ('segpar', 'relationships', 'landnos')].rename({'relationships' : 'rels'}, axis=1)

# Split the land numbers string into list
parcels_land.landnos = parcels_land.landnos.apply(lambda v: [int(x) for x in v.split(';')])

# Explode on the land numbers column (this will create duplicate parcel segpar records)
parcels_land = parcels_land.explode('landnos', ignore_index=True).rename({'landnos' : 'landno'}, axis=1)

# Inspect
parcels_land.head()

''' 
----------------------------------
Land Aggregates

Identify the land aggregates for each parcels
Land aggregates are where there are multiple council land parcels linked to a DCDB parcel
----------------------------------
'''

# Select only those parcels that have more than one relation
parcels_landagg = parcels_land.loc[parcels_land['rels'].astype(int) > 1]

# Group the land parcels by DCDB segpar
land_agg = parcels_landagg.loc[:, ('segpar', 'landno')].groupby(by='segpar').count().rename({'landno' : 'aggct'}, axis=1).reset_index()

# Create a unique aggregate id for the matched building unit plans
land_agg['aggid'] = land_agg.apply(lambda row: f'LND{row.name + 1:06}', axis=1)

# Joint the agg ids back onto the parcles
parcels_landagg = pd.merge(parcels_landagg, land_agg, on='segpar', how='left').drop(['rels'], axis=1)

# Inspect
parcels_landagg.head()

''' 
----------------------------------
BUP Aggregates

Identify the building unit plan aggregates for each parcel
----------------------------------
'''

# Import the building unit plan table
bup = pd.DataFrame.spatial.from_table(os.path.join(input_gdb_path, 'INPUT_TB_DCDB_BuildingUnitPlans')).drop(['OBJECTID'], axis=1)

# Join the BUP table to parcles and filter out non-joins
parcels_bup = pd.merge(parcels[['segpar', 'lotplan']], bup, on='lotplan', how='left').dropna()

# Create the building unit plan to parcels aggregate table
bup_agg = pd.merge(
    parcels_bup[['bup_plan', 'segpar']].groupby(by='bup_plan').count(),
    parcels_bup[['bup_plan', 'lot_area_am']].groupby(by='bup_plan').sum(),
    on='bup_plan',
    how='left'
).rename({'segpar' : 'aggct'}, axis=1).reset_index()

# Create a unique aggregate id for the matched building unit plans
bup_agg['aggid'] = bup_agg.apply(lambda row: f'BUP{row.name + 1:06}', axis=1)

# Join the building unit plan aggregates back onto the parcles layer
parcels_bup = pd.merge(parcels_bup[['segpar', 'bup_plan']].drop_duplicates(), bup_agg, on='bup_plan', how='left')

# Inspect
parcels_bup.head()

''' 
----------------------------------
Named Aggregates

Identify the named aggregates for each parcel (based on the DCDB alias name)
----------------------------------
'''

# Get the named exclusions from the input vars table as list
alias_excl = pd.read_excel(vars_xlsx, 'DCDB_AliasExclude')['alias_name'].tolist()

# Extract a subset of the parcels table for only those with an alias name
parcels_named = parcels[['segpar', 'alias_name']].dropna()

# Drop those rows that have an excluded alias name
parcels_named = parcels_named.loc[~parcels_named['alias_name'].isin(alias_excl)]

# Group the parcels by name
named_agg = parcels_named.groupby(by='alias_name').count().rename({'segpar' : 'aggct'}, axis=1)

# Drop all the rows with a count less than 2
named_agg = named_agg.loc[named_agg['aggct'] > 1].reset_index()

# Create a unique aggregate id for the matched alias names
named_agg['aggid'] = named_agg.apply(lambda row: f'NME{row.name + 1:06}', axis=1)

# Join the named aggregates back onto the parcels (dropping non-joins)
parcels_named = pd.merge(parcels_named[['segpar', 'alias_name']].drop_duplicates(), named_agg, on='alias_name', how='left').dropna().reset_index(drop=True)

# Set the agg id and count columns to integer
parcels_named['aggct'] = parcels_named['aggct'].astype('Int64')

# Inspect
parcels_named.head()


''' 
----------------------------------
Holding Aggregates

Identify the holding aggregates for each parcel
----------------------------------
'''

# Property holdings
land_holdings = land[['landno', 'propno']].drop_duplicates()

# Group the parcels by holding
holdings_agg = land_holdings.groupby(by='propno').count().rename({'landno' : 'aggct'}, axis=1)

# Drop all the rows with a count less than 2
holdings_agg = holdings_agg.loc[holdings_agg['aggct'] > 1].reset_index()

# Create a unique aggregate id for the property holdings
holdings_agg['aggid'] = holdings_agg.apply(lambda row: f'HLD{row.name + 1:06}', axis=1)

# Join the holding aggregates back onto the land holdings (dropping non-joins)
holdings_agg = pd.merge(land_holdings, holdings_agg, on='propno', how='left').dropna().reset_index(drop=True)

# Set the agg count columns to integer
holdings_agg['aggct'] = holdings_agg['aggct'].astype('Int64')

# Get a copy of the parcels table
parcels_holding = parcels_land.loc[:, ('segpar', 'landno')]

# Join the holding aggregate id and count onto the parcel records
parcels_holding = pd.merge(parcels_holding, holdings_agg, on='landno', how='left').dropna().reset_index(drop=True)

# Set the joined columns to integer
parcels_holding['propno'] = parcels_holding['propno'].astype('Int64')
parcels_holding['aggct'] = parcels_holding['aggct'].astype('Int64')

# Inspect
parcels_holding.loc[parcels_holding['aggid'] == 'HLD000525']

''' 
----------------------------------
Address

Identify the address aggregates for each parcel

NOTE: Address aggregates are now redundant and the outputs of this cell will not be used
----------------------------------
'''

# Addresses
land_address = land.loc[:, ('landno', 'unit_nos', 'house_nos', 'street', 'locality', 'postcode')].rename({
    'unit_nos' : 'unit', 
    'house_nos' : 'house',
}, axis=1)

# List of address columns
address_cols = ['house_from', 'house_to', 'street', 'locality', 'postcode']

# Function to get house from and to details and strip the non numeric characters
def houseFromTo(n, c='-'):
    if n:
        ft = n.split(c)
        f = int(re.sub(r'[^0-9.]', '', ft[0]))
        t = int(re.sub(r'[^0-9.]', '', ft[len(ft) - 1]))
        return[f, t]
    else:
        return [0, 0]

# Create a house from and to column using the house number and remove all non integer characters
land_address.loc[:,'house_from'] = land_address.loc[:,'house'].apply(lambda x: houseFromTo(x)[0])
land_address.loc[:,'house_to'] = land_address.loc[:,'house'].apply(lambda x: houseFromTo(x)[1])

# Create a new concatenated address column
land_address['address'] = \
    land_address['house_from'].astype(str) + '-' + \
    land_address['house_from'].astype(str) + ', ' + \
    land_address['street'].astype(str) + ', ' + \
    land_address['locality'].astype(str) + ' ' + \
    land_address['postcode'].astype(str)

# Group the land parcels by address
address_agg = land_address.loc[:, ('landno', 'address')].groupby(by='address').count().rename({'landno' : 'aggct'}, axis=1)

# Drop all the rows with a count less than 2
address_agg = address_agg.loc[address_agg['aggct'] > 1].reset_index()

# Create a unique aggregate id for the property address aggregates
address_agg['aggid'] = address_agg.apply(lambda row: f'ADD{row.name + 1:06}', axis=1)

# Join the address aggregates back onto the land address table (dropping non-joins)
address_agg = pd.merge(land_address, address_agg, on='address', how='left').dropna(subset=['aggct', 'aggid']).reset_index(drop=True)

# Set the agg id column to integer
address_agg['aggct'] = address_agg['aggct'].astype('Int64')

# Get a copy of the parcels table
parcels_address = parcels_land.loc[:, ('segpar', 'landno')]

# Join the address aggregate id and count onto the parcel records
parcels_address = pd.merge(parcels_address, address_agg, on='landno', how='left').dropna(subset=['aggct', 'aggid']).reset_index(drop=True)

# Inspect
parcels_address.loc[parcels_address['aggid'] == 'ADD000653'].sort_values(by='segpar')


''' 
----------------------------------
Import Shared Boundaries
----------------------------------
'''

# Import the DCDB Boundaries feature class as a dataframe
parcel_bounds = pd.DataFrame.spatial.from_featureclass(os.path.join(input_gdb_path, 'INPUT_FC_DCDB_SharedBoundaries')).drop(['OBJECTID', 'SHAPE'], axis=1)

# Insepct
parcel_bounds.head()

''' 
----------------------------------
Common Property

Identify the common property aggregates for each parcel
----------------------------------
'''

# Get property use table
propuse = pd.read_excel(vars_xlsx, 'PropertyUse_to_DevGroup')

# Merge property use tags onto council land
land_propuse = pd.merge(land.loc[:, ('landno', 'propno', 'prop_use')], propuse.loc[:, ('prop_use', 'tag')], on='prop_use', how='left').dropna(subset=['tag'])

# Reset property number to be an integer
land_propuse['propno'] = land_propuse['propno'].astype('Int64')

# Link the property use to parcels via the parcels to council land mapping table
parcels_propuse = pd.merge(parcels_land, land_propuse, on='landno', how='left')

# Select all common properties from the data frame
common_props = parcels_propuse.loc[parcels_propuse['tag'] == 'Common'].copy()

# Get all parcels that share a boundary with the common properties
common_props['touching_parcels'] = common_props['segpar'].apply(lambda segpar: set(parcel_bounds.loc[parcel_bounds['shared_segpar'] == segpar]['segpar'].to_list()))

# --------------------------
# BUP

# Get the bup aggerate id linked to the common property
common_props['bup_aggid'] = common_props['segpar'].apply(lambda segpar: set(parcels_bup.loc[parcels_bup['segpar'] == segpar]['aggid'].to_list()))

# Get the linked parcels for the bup aggregates
common_props['bup_parcels'] = common_props['bup_aggid'].apply(lambda aggids : \
    set().union(*[set(parcels_bup.loc[parcels_bup['aggid'] == aggid]['segpar'].to_list()) for aggid in aggids if len(aggid) > 0]) \
)

# --------------------------
# Named

# Get the named aggerate id for the common property
common_props['named_aggid'] = common_props['segpar'].apply(lambda segpar: set(parcels_named.loc[parcels_named['segpar'] == segpar]['aggid'].to_list()))

# Get the linked parcels for the named aggregates
common_props['named_parcels'] = common_props['named_aggid'].apply(lambda aggids : \
    set().union(*[set(parcels_named.loc[parcels_named['aggid'] == aggid]['segpar'].to_list()) for aggid in aggids if len(aggid) > 0]) \
)

# --------------------------
# Holding

# Get the holding aggerate id linked to the common property
common_props['holding_aggid'] = common_props['segpar'].apply(lambda segpar: set(parcels_holding.loc[parcels_holding['segpar'] == segpar]['aggid'].to_list()))

# Get the linked parcels for the holding aggregates
common_props['holding_parcels'] = common_props['holding_aggid'].apply(lambda aggids : \
    set().union(*[set(parcels_holding.loc[parcels_holding['aggid'] == aggid]['segpar'].to_list()) for aggid in aggids if len(aggid) > 0]) \
)

# --------------------------
# Land

# Get the address aggerate id for the common property
common_props['land_aggid'] = common_props['segpar'].apply(lambda segpar: set(parcels_landagg.loc[parcels_landagg['segpar'] == segpar]['aggid'].to_list()))

# Get the linked parcels for the address aggregates
common_props['land_parcels'] = common_props['land_aggid'].apply(lambda aggids : \
    set().union(*[set(parcels_landagg.loc[parcels_landagg['aggid'] == aggid]['segpar'].to_list()) for aggid in aggids if len(aggid) > 0]) \
)

# --------------------------

# Combile all linked aggregate parcels into single field
common_props['agg_parcels'] = common_props.apply(lambda row: set().union(*[row.bup_parcels, row.named_parcels, row.holding_parcels, row.land_parcels]), axis=1)

# Drop redundant columns
common_props = common_props.drop(['bup_aggid', 'bup_parcels', 'named_aggid', 'named_parcels', 'holding_aggid', 'holding_parcels', 'land_aggid', 'land_parcels'], axis=1)

# Drop any of the touching parcels that do not share a common aggregate
common_props['linked_parcels'] = common_props.apply(lambda row: row.touching_parcels.intersection(row.agg_parcels), axis=1)

# Keep necessary columns
common_props = common_props.loc[:, ('segpar', 'linked_parcels')]

# --------------------------

# Inset the segpar into the linked parcels for completeness
common_props['linked_parcels'] = common_props.apply(lambda row: [*row.linked_parcels.union([row.segpar])], axis=1)

# Sort values
common_props = common_props.sort_values(by='segpar')

# Inspect
common_props.head(10)

''' 
----------------------------------
Compile common property aggregates
----------------------------------
'''
# --------------------------------

'''
The following function moves through the grouped parcels and creates a list
of final aggregates that account for when a parcel falls within multiple groups.
In this case all of the groups and parcels that are related by one or many 
shared parcles are pulled together to form a final 'aggregate of aggregates' 
grouping. This process handles duplicate parcel 'segpar' identifiers by making 
use of python sets which cannot contain duplicate entries.
'''

def compileAggregates(series):

    # Define a container for the list of final aggregates
    final_aggregates = []

    # First pass run through of aggregate groups
    for id, group in series.items():
        
        g = set(group)

        for i, agg in enumerate(final_aggregates):
            if len(agg.intersection(g)) > 0:
                final_aggregates[i] = agg.union(g)
                break
        else:
            final_aggregates.append(g)

    # A procedure to contain temporary aggregate groupings
    temp_aggregates = []

    # Second pass run through of the aggregates
    while True:

        for agg in final_aggregates:
            for i, temp_agg in enumerate(temp_aggregates):
                if len(temp_agg.intersection(agg)) > 0:
                    temp_aggregates[i] = temp_agg.union(agg)
                    break
            else:
                temp_aggregates.append(agg)

        if len(temp_aggregates) == len(final_aggregates):
            break
        else:
            final_aggregates = temp_aggregates[:]
            temp_aggregates = []

    return final_aggregates

# Group parcles by aggregate id (creates series)
parcels_common_comp = compileAggregates(common_props['linked_parcels'])

# Create final dataframe from the processed sets
parcels_common = pd.DataFrame(list(zip(
        [f'CMN{i + 1:06}' for i, v in enumerate(parcels_common_comp)], # Create aggregate IDs
        [[*x] for x in parcels_common_comp] # Convert segpar ids list of sets to list of lists
    )), columns=['aggid', 'segpar'])

# Explode on segpar
parcels_common = parcels_common.explode('segpar').reset_index(drop=True)

# Inspect
parcels_common.head()

''' 
----------------------------------
Create a mapping of all aggregates for use down stream
----------------------------------
'''

# Concatenate all aggregate dataframes
all_aggregates = pd.concat([
    parcels_bup.loc[:,('segpar','aggid')],
    parcels_named.loc[:,('segpar','aggid')],
    parcels_holding.loc[:,('segpar','aggid')],
    parcels_landagg.loc[:,('segpar','aggid')],
    parcels_common.loc[:, ('segpar', 'aggid')]
])

# Aggregate type helper function
def getAggType(idstr):

    if idstr[:3] == 'LND':
        return 'land'
    if idstr[:3] == 'BUP':
        return 'bup'
    if idstr[:3] == 'HLD':
        return 'holding'
    if idstr[:3] == 'NME':
        return 'named'
    if idstr[:3] == 'CMN':
        return 'common'
    else:
        return None

# Extract the aggregate type by id
all_aggregates['aggtype'] = all_aggregates.loc[:,('aggid')].apply(lambda x: getAggType(x))

# Inspect
all_aggregates.tail()

''' 
----------------------------------
Process all aggregates and allocate a single aggregate id for each parcel
----------------------------------
'''

# Group parcles by aggregate id (creates series)
agg_groups = all_aggregates.groupby(by='aggid')['segpar'].apply(list)

# Define a container for the list of final aggregates
final_aggregates = compileAggregates(agg_groups)

# Create an aggregates dataframe from the processed sets
aggregates = pd.DataFrame(list(zip(
        [f'AGG{i + 1:06}' for i, v in enumerate(final_aggregates)], # Create aggregate IDs
        [[*x] for x in final_aggregates] # Convert segpar ids list of sets to list of lists
    )), columns=['aggid', 'segpar'])

# Inspect
aggregates.head()

''' 
----------------------------------
Filter aggregates by type and output final list
----------------------------------
'''

# Get parcel list (as set) for each aggregate type
lnd_ls = {*all_aggregates.loc[all_aggregates['aggtype'] == 'land'].loc[:,('segpar')].to_list()}
bup_ls = {*all_aggregates.loc[all_aggregates['aggtype'] == 'bup'].loc[:,('segpar')].to_list()}
hld_ls = {*all_aggregates.loc[all_aggregates['aggtype'] == 'holding'].loc[:,('segpar')].to_list()}
nme_ls = {*all_aggregates.loc[all_aggregates['aggtype'] == 'name'].loc[:,('segpar')].to_list()}
cmn_ls = {*all_aggregates.loc[all_aggregates['aggtype'] == 'common'].loc[:,('segpar')].to_list()}

# Get parcel list for pure named aggregates (name only - no other linked aggregates)
pure_nme_ls = nme_ls.difference(set().union(*[lnd_ls, bup_ls, hld_ls, cmn_ls]))

# Explode the aggregates by parcel
parcels_agg = aggregates.explode('segpar', ignore_index=True)

# Discard pure named aggregates
parcels_agg = parcels_agg[~parcels_agg['segpar'].isin(set().union(*[pure_nme_ls]))]

# Inspect
parcels_agg.head()

''' 
----------------------------------
Prepare aggregate details table
----------------------------------
'''

# Join the all aggregates table to parcels land on segpar
aggregate_details = pd.merge(parcels_land, all_aggregates, on='segpar', how='left').dropna(subset=['aggid']).drop(['rels', 'aggid'], axis=1)

# Join in the parcels agg table to filter out redundant aggregates
aggregate_details = pd.merge(aggregate_details, parcels_agg, on='segpar', how='left').dropna(subset=['aggid'])

# Inspect
aggregate_details.head()

''' 
----------------------------------
Export tables to geodatabase
----------------------------------
'''

# Write the aggregate details table back into the geodatabase
aggregate_details.spatial.to_table(os.path.join(input_gdb_path, 'GEN_TB_DCDB_AggregateDetails'), overwrite=True)

# Write the final aggregates back into the geodatabase
parcels_agg.spatial.to_table(os.path.join(input_gdb_path, 'GEN_TB_DCDB_Aggregates'), overwrite=True)
