==> notebooks/P5_allocate_nonres_growth.ipynb <==
# Establish connections to the input db and variables spreadsheets

# Import modules
import os, re, math
import arcpy
import pandas as pd
import numpy as np
from arcgis.features import GeoAccessor, GeoSeriesAccessor
from dotenv import load_dotenv

# Lot the .env variables
load_dotenv()

# Set the data paths
inputs_path = os.getenv('J111_INPUTS')
outputs_path = os.getenv('J111_OUTPUTS')
spatial_path = os.getenv('J111_SPATIAL')
env_path = os.getenv('J111_ENV')

# Set path to the input excel workbooks
da_xlsx = os.path.join(inputs_path, 'INPUT_DevelopmentApplications.xlsx')
proj_xlsx = os.path.join(inputs_path, 'INPUT_InterimProjections_NonResidential.xlsx')

# Get full path of Input GDBs
pscap_gdb_path = os.path.join(spatial_path, 'redland_gam_ps_capacity.gdb')
curdev_gdb_path = os.path.join(spatial_path, 'redland_gam_current_dev.gdb')
alloc_gdb_path = os.path.join(spatial_path, 'redland_gam_alloc.gdb')

# Set arcpy workspace
arcpy.env.workspace = os.path.join(env_path, 'J111_redland_gam.gdb')

# *Variables*

''' 
----------------------------------
Define allocation variables
----------------------------------
'''

years = [2026, 2031, 2036, 2041, 2046, 2051]
nonres_cols = [
    'commercial',
    'community',
    'education',
    'health',
    'industrial',
    'other',
    'retail',
]

# *Development Assessment Data*

''' 
----------------------------------
Import development assessment data
----------------------------------
'''

# Import development assessment data table
da_data = pd.read_excel(da_xlsx, 'DevelopmentApplications').drop(['app_detail', 'address', 'decision'], axis=1)

# Inspect
da_data.head()

''' 
----------------------------------
Import and prepare the parcels to council land relational dataframe
----------------------------------
'''

# Import the Council Property to DCDB Base lots relational list
parcels_to_land = pd.DataFrame.spatial.from_table(os.path.join(curdev_gdb_path, 'INPUT_TB_DCDB_CouncilPropertyRel')).drop(['OBJECTID'], axis=1)

# Select only the required column
parcels_land = parcels_to_land.loc[:, ('segpar', 'landnos')]

# Split the land numbers string into list
parcels_land.landnos = parcels_land.landnos.apply(lambda v: [int(x) for x in v.split(';')])

# Explode on the land numbers column (this will create duplicate parcel segpar records)
parcels_land = parcels_land.explode('landnos', ignore_index=True).rename({'landnos' : 'landno'}, axis=1)

# Inspect
parcels_land.head()

''' 
----------------------------------
Relate DA data to land parcels
----------------------------------
'''

# Get subset of the DA data
da_land = da_data.loc[:, ('appid', 'landnos')]

# Set and explode list on the landnos
da_land['landnos'] = da_land['landnos'].astype(str) + ';'
da_land['landnos'] = da_land['landnos'].str.split(';')
da_land['landnos'] = da_land.apply(lambda row: list(filter(None, row.landnos)), axis=1)
da_land = da_land.explode('landnos').rename({'landnos' : 'landno'}, axis=1)
da_land['landno'] = da_land.apply(lambda row: int(str(row.landno)), axis=1)
da_land['landno'] = da_land['landno'].astype('Int64')

# Get the segpar for each landno
da_land = pd.merge(da_land, parcels_land, on='landno', how='left')

# Import the aggregate details table from geodatabase
agg_details = pd.DataFrame.spatial.from_table(os.path.join(curdev_gdb_path, 'GEN_TB_DCDB_AggregateDetails')).drop(['OBJECTID'], axis=1)

# Drop duplicates from the aggregate details
agg_details = agg_details.drop(['aggtype', 'segpar'], axis=1).drop_duplicates().sort_values(by='aggid').reset_index(drop=True)

# Join aggid to DA data via landno
da_land = pd.merge(da_land, agg_details, on='landno', how='left')

# Split DA data into those with an aggid and those without
da_land_agg = da_land.query("aggid == aggid").drop(['landno', 'segpar'], axis=1).drop_duplicates().reset_index(drop=True)
da_land_other = da_land.query("aggid != aggid").drop(['landno', 'aggid'], axis=1).drop_duplicates().reset_index(drop=True)

# Inspect
da_land_agg.head()

# Import propbase
propbase = pd.DataFrame.spatial.from_featureclass(os.path.join(curdev_gdb_path, 'GEN_FC_PropertyBase')).drop(['OBJECTID', 'SHAPE'], axis=1)

# Inspect
propbase.head()

''' 
----------------------------------
Relate DA data to propbase
----------------------------------
'''

# Split the propbase into aggregates and otherwise
pb_agg = propbase.query("aggid == aggid").drop(['segpar', 'sa2_name'], axis=1).drop_duplicates().reset_index(drop=True)
pb_other = propbase.query("aggid != aggid").drop(['aggid', 'sa2_name'], axis=1).drop_duplicates().reset_index(drop=True)
pb_other['segpar'] = pb_other['segpar'].astype('Int64')

# Join the propbase data onto the respective dataframes and concatenate
pb_da = pd.concat([
    pd.merge(da_land_agg, pb_agg, on='aggid', how='left').drop(['aggid'], axis=1),
    pd.merge(da_land_other, pb_other, on='segpar', how='left').drop(['segpar'], axis=1)
]).sort_values('pbno').reset_index(drop=True)

# Join DA data back onto related table
pb_da = pd.merge(pb_da, da_data.drop(['landnos'], axis=1), on='appid', how='left')

# ----------------------------------
# IMPORTANT !!!

# If there are multiple applications on a property take the most recent
pb_da = pb_da.sort_values('decision_date').drop_duplicates('pbno', keep='last').drop(['decision_date'], axis=1).reset_index(drop=True)

# ----------------------------------

# Create a subset of the table and allocate development data where situated across multiple properties
pb_da_sub = pb_da.drop(['pbno'], axis=1)
pb_da_sub['props'] = 1
ps_da_dub_g1 = pb_da_sub.loc[:, ('appid', 'props')].groupby('appid').agg('count').reset_index()
ps_da_dub_g2 = pb_da_sub.groupby('appid').agg('max').drop(['props'], axis=1).reset_index()
pb_da_sub = pd.merge(ps_da_dub_g1, ps_da_dub_g2, on='appid', how='left')

# Force all values on development columns as floats
pb_da_sub.iloc[:,2:] = pb_da_sub.iloc[:,2:].astype(float)

# Divide the development columns by the number to props to allocate
pb_da_sub.iloc[:,2:] = pb_da_sub.iloc[:,2:].div(pb_da_sub.props, axis=0)

# Drop columns
pb_da_sub = pb_da_sub.drop(['props'], axis=1)

# ----------------------------------

# Merge cleaned data back onto the probase-da table
pb_da = pd.merge(pb_da.iloc[:,:2], pb_da_sub, on='appid', how='left')

# Split the residential and non-residential components into separate dataframes
pb_da_res = pb_da.drop([col for col in pb_da.columns if col not in ['app_id', 'pbno', 'att_dwl', 'det_dwl']], axis=1)
pb_da_nonres = pb_da.drop(['det_dwl', 'att_dwl'], axis=1)

# Inspect
pb_da_res.head()

# *Current Development Statistics*

''' 
----------------------------------
Import and prepare current development stats
----------------------------------
'''

# Import current development metrics from geodatabase
pb_curdev_lyr = pd.DataFrame.spatial.from_featureclass(os.path.join(curdev_gdb_path, 'OUT_FC_PropertyBase_CurrentDev')).drop(['OBJECTID'], axis=1)

# Drop residential columns
pb_curdev = pb_curdev_lyr.copy().drop(['SHAPE', 'prop_use', 'det_dwl', 'att_dwl', 'accom_long'], axis=1)

# Summarise current non-residential development into broader development groups

pb_curdev['commercial'] = pb_curdev['office']
pb_curdev['community'] = pb_curdev['emer_serv'] + pb_curdev['pl_of_ass']
pb_curdev['education'] = pb_curdev['childcare'] + pb_curdev['education']
pb_curdev['health'] = pb_curdev['care_accom'] + pb_curdev['hospital'] + pb_curdev['health_serv']
pb_curdev['industrial'] = pb_curdev['heavy_ind'] + pb_curdev['light_ind']
pb_curdev['other'] = pb_curdev['accom_short'] + pb_curdev['other'] + pb_curdev['rural'] + pb_curdev['indoor_ent'] + pb_curdev['indoor_rec']
pb_curdev['retail'] = pb_curdev['bulk_goods'] + pb_curdev['food_drink'] + pb_curdev['retail']

# Drop redundant non-residential columns
drop_nonres_col = [
    'office',
    'accom_short',
    'emer_serv',
    'pl_of_ass',
    'childcare',
    'care_accom',
    'hospital',
    'health_serv',
    'heavy_ind',
    'light_ind',
    'rural',
    'bulk_goods',
    'food_drink',
    'indoor_ent',
    'indoor_rec'
]
pb_curdev = pb_curdev.drop(drop_nonres_col, axis=1)

# Get total non-residential gfa by summing nonres columns
pb_curdev['nonres_gfa'] = pb_curdev.loc[:, tuple(nonres_cols)].sum(axis=1)

# Re-name columns to identify as 2021 
pb_curdev_nonres = pb_curdev.rename({col: col + '_2021' for col in pb_curdev.columns if col not in ['pbno', 'sa2_name']}, axis=1)

# Inspect
pb_curdev_nonres.head()

# *Development Capacity Statistics*

''' 
----------------------------------
Import and prepare development capacity table
----------------------------------
'''

# Import development table from geodatabase
pb_devcap_lyr = pd.DataFrame.spatial.from_featureclass(os.path.join(pscap_gdb_path, 'OUT_FC_PropertyBase_DevCapacity')).drop(['OBJECTID'], axis=1)

# Copy the layer into a dataframe
pb_devcap = pb_devcap_lyr.copy().drop(['zone', 'property_area', 'eda', 'SHAPE'], axis=1)

# Get total non-residential gfa by summing nonres columns
pb_devcap['nonres_gfa'] = pb_devcap.loc[:, tuple(nonres_cols)].sum(axis=1)

# Re-name columns to identify as planned capacity
pb_devcap = pb_devcap.rename({col: col + '_cap' for col in pb_devcap.columns if col not in ['pbno', 'sa2_name', 'res_conv', 'nonres_conv']}, axis=1)

# Select only those columns that will convert
pb_devcap['conv'] = pb_devcap['res_conv'] + pb_devcap['nonres_conv']
pb_devcap = pb_devcap.query("conv > 0").drop(['conv'], axis=1)

# Split the residential and non-residential conversions into separate dataframes
pb_devcap_nonres = pb_devcap.query("nonres_conv > 0").drop(['res_conv', 'det_dwl_cap', 'att_dwl_cap'], axis=1).reset_index(drop=True)

# Inspect
pb_devcap_nonres.head()

# *Allocation functions*

# Helper function for initializing spare capacity
def InitSpareCapacity(prop, dev_type, proj_year):

    dev_year = prop[f'{dev_type}{proj_year}']
    dev_cap = prop[f'{dev_type}cap']

    if dev_year and dev_cap > 0:
        if dev_cap - dev_year > 0:
            return dev_cap - dev_year
        else:
            return 0
    else:
        return dev_cap

# Helper function for setting remaining dev values in projection year (accounting for existing data)
def SetYear(prop, dev_type, proj_year, base_year):

    dev_year = prop[f'{dev_type}{proj_year}']
    dev_prev = prop[f'{dev_type}{base_year}']

    if dev_year == 0:
        return dev_prev
    else:
        return dev_year

# Helper function to set non-residential DA fa
def SetDevAssGfa(prop, dev_type, base_year):

    da_gfa = prop[f'{dev_type}da']
    base_gfa = prop[f'{dev_type}{base_year}']

    if da_gfa == 0.0:
        return 0.0
    elif da_gfa <= base_gfa:
        return base_gfa + da_gfa
    else:
        return da_gfa


# *Non-residential growth allocation process*

''' 
----------------------------------
Handle DA non-residential allocation in the 2026 horizon
----------------------------------
'''
# Prepare initial dataframe
pb_alloc_nonres_da = pb_curdev_nonres.set_index('pbno')

# Set new columns to 0
for col in nonres_cols:
    pb_alloc_nonres_da[f'{col}_2026'] = 0.0
    pb_alloc_nonres_da[f'{col}_cap'] = 0.0
    pb_alloc_nonres_da[f'{col}_da'] = 0.0

pb_alloc_nonres_da['nonres_conv'] = 0

# Set 2026 columns with da data where matched on pbno
pb_alloc_nonres_da.update(pb_da_nonres.rename({col: col + '_da' for col in pb_da_nonres.columns if col != 'pbno'}, axis=1).set_index('pbno'))

# Set capacity columns where matched on pbno
pb_alloc_nonres_da.update(pb_devcap_nonres.drop(['sa2_name'], axis=1).set_index('pbno'))

# Update the gfa numbers across all columns
for col in nonres_cols:
    
    # Set the GFA of 2026 to be either the DA GFA if it is greater or the DA GFA + existing GFA if it is lesser
    pb_alloc_nonres_da[f'{col}_2026'] = pb_alloc_nonres_da.apply(lambda row: SetDevAssGfa(row, f'{col}_', 2021), axis=1)

    # If there is da data for 2026 subtract this from the capacity
    pb_alloc_nonres_da[f'{col}_cap'] = pb_alloc_nonres_da.apply(lambda row: max(0.0, row[f'{col}_cap'] - row[f'{col}_2026']) if row[f'{col}_2026'] > 0 else row[f'{col}_cap'],axis=1)

# Set the projection area based on SA2
pb_alloc_nonres_da['area_name'] = pb_alloc_nonres_da.apply(lambda row: 'Balance' if row['sa2_name'] != 'Redland Islands' else row['sa2_name'], axis=1)

# Inspect
pb_alloc_nonres_da.sort_values('nonres_conv', ascending=False).head()

''' 
----------------------------------
Import QGSO non-residential projections
----------------------------------
'''

# Import detached dwelling projections
nonres_proj = pd.read_excel(proj_xlsx, 'OUT_NonResProjections')

# Select only the growth columns
nonres_proj = nonres_proj.loc[:, ~nonres_proj.columns.str.startswith('proj_')]

# Create a dictionary that holds a single dataframe for each non-residential category
nonres_projs = {}

# Loop through non-residential development categories
for category in nonres_cols:

    # Select the rows for this category
    category_df = nonres_proj.query("sector == @category").drop(['sector'], axis=1).reset_index(drop=True)

    # Rename the growth columns accordingly
    category_df = category_df.rename({col: f'{category}_' + col.lstrip('growth_') for col in category_df.columns if col != 'area_name'}, axis=1)

    # Push dataframe into dictionary
    nonres_projs[category] = category_df.set_index('area_name')

# Inspect
nonres_projs['industrial']

''' 
----------------------------------
Correct the non-residential growth projections to account for the already allocated DA data
----------------------------------
'''

# Summarise the DA data by projection area
da_nonres_sum = pb_alloc_nonres_da.drop(['sa2_name'], axis=1).reset_index(drop=True).loc[:, ('area_name',) + tuple([f'{col}_da' for col in nonres_cols])].groupby('area_name').agg('sum')

# Get a copy of the non-residential projections dataframes
nonres_projs_corr = nonres_projs

for cat, proj in nonres_projs_corr.items():
        
    carry = da_nonres_sum[f'{cat}_da']

    for year in years:

        sub = proj[f'{cat}_{year}'].sub(carry)
        proj[f'{cat}_{year}'] = sub.apply(lambda x: 0 if x <= 0 else x)
        
        carry = sub.apply(lambda x: abs(x) if x < 0 else 0)

# Inspect
nonres_projs_corr['industrial']

''' 
----------------------------------
Allocate non-residential growth projections per property based on remaining capacity
----------------------------------
'''

# Prepare initial dataframe
pb_alloc_nonres = pb_alloc_nonres_da.copy()

# Drop DA GFA columns
pb_alloc_nonres = pb_alloc_nonres.drop([f'{col}_da' for col in nonres_cols], axis=1)

# Set the projection area
pb_alloc_nonres['area_name'] = pb_alloc_nonres_da.apply(lambda row: 'Balance' if row['sa2_name'] != 'Redland Islands' else row['sa2_name'], axis=1)

# Set spare capacity
for col in nonres_cols:
    pb_alloc_nonres[f'{col}_spare'] = pb_alloc_nonres.apply(lambda row: InitSpareCapacity(row, f'{col}_', 2026), axis=1)

# Set first projeciton year (2026)
for col in nonres_cols:
    pb_alloc_nonres[f'{col}_{2026}'] = pb_alloc_nonres.apply(lambda row: SetYear(row, f'{col}_', 2026, 2021), axis=1)

# Drop residential conversion column as this is now redundant
pb_alloc_nonres = pb_alloc_nonres.drop(['nonres_conv'], axis=1)

# ----------------------------------
# Set empty years

for col in nonres_cols:

    for year in years[1:]:

        pb_alloc_nonres[f'{col}_{year}'] = 0.0

# ----------------------------------
# Set column order
nonres_col_order = ['area_name', 'sa2_name']

for cat in nonres_cols:

    nonres_col_order += [f'{cat}_{year}' for year in [2021] + years] + [f'{cat}_spare', f'{cat}_cap']

pb_alloc_nonres = pb_alloc_nonres[nonres_col_order]

# ----------------------------------
# ALLOCATION LOOP

# For each of the projection categories
for cat, projdf in nonres_projs_corr.items():

    # For each projection year
    for i, year in enumerate(years):

        # Set the non-residential development column names
        dev_year = f'{cat}_{year}'
        dev_prev = f'{cat}_{year - 5}'
        dev_cap = f'{cat}_cap'
        dev_spare = f'{cat}_spare'

        # Get the summary statistics for the allocation
        year_spare = pb_alloc_nonres.reset_index().loc[:, ('area_name', dev_spare)].groupby('area_name').sum().squeeze()
        year_proj = projdf.reset_index().loc[:, ('area_name', dev_year)].set_index('area_name').squeeze()

        # Calculate the properties contribution to the spare capacity within the area based on its development capacity
        pb_alloc_nonres['contrib'] = pb_alloc_nonres.apply(lambda row: row[dev_cap] / year_spare[f'{row.area_name}'] if row[dev_spare] > 0 else 0, axis=1)

        # For each projection area grab only those properties with a contribution
        for area_name, proj in year_proj.iteritems():
            
            # Get subset of the allocation dataframe
            pb_alloc_sub = pb_alloc_nonres.query("area_name == @area_name").reset_index().loc[:, ('pbno', f'{dev_prev}', f'{dev_year}', f'{dev_spare}', 'contrib')]

            # Sort the subset from the top down
            pb_alloc_sub = pb_alloc_sub.sort_values('contrib', ascending=False)

            # Loop through the subset and allocate growth carrying over remainder when allocating
            alloc = proj

            for index, prop in pb_alloc_sub.iterrows():
                
                prop_dev_cur = prop[f'{dev_year}']
                prop_dev_prev = prop[f'{dev_prev}']
                prop_dev_spare = prop[f'{dev_spare}']

                if prop_dev_cur == 0.0 and prop_dev_spare:
                    
                    if alloc >= prop_dev_spare:
                        
                        prop_dev_frac = max(0.0, prop.contrib * proj)

                        if prop_dev_spare <= 100.0:
                            
                            prop_new_dev = prop_dev_spare

                        elif prop_dev_frac >= 100.0:

                            if prop_dev_frac > prop_dev_spare:

                                prop_new_dev = prop_dev_spare

                            else:

                                prop_new_dev = prop_dev_frac

                        else:

                            continue

                    else:

                        # Set the new development as the remaining allocation
                        prop_new_dev = alloc

                    # Update the development column in the subset
                    pb_alloc_sub.at[index, f'{dev_year}'] = prop_new_dev + prop_dev_prev

                    # Update the spare capacity column in the subset
                    pb_alloc_sub.at[index, f'{dev_spare}'] = prop_dev_spare - prop_new_dev

                    # Remove new development from the overall allocation
                    alloc -= prop_new_dev

                if alloc <= 0:

                    break

            # Drop extraneous columns from the subset
            pb_alloc_sub = pb_alloc_sub.drop(['contrib'], axis=1).set_index('pbno')

            # Update the master dataframe with the updated columns
            pb_alloc_nonres.update(pb_alloc_sub)

        # Set remaining null values in projection year to be that of year before (for 2031 onwards)
        if year >= 2031:
            pb_alloc_nonres[f'{dev_year}'] = pb_alloc_nonres.apply(lambda row: SetYear(row, f'{cat}_', year, years[i - 1]), axis=1)

# ----------------------------------
# Drop extraneous
pb_alloc_nonres = pb_alloc_nonres.drop(['area_name', 'contrib'], axis=1)

# Inspect
insp = 'industrial'
insp_cols = ['sa2_name'] + [f'{insp}_{year}' for year in [2021] + years] + [f'{insp}_spare', f'{insp}_cap']
pb_alloc_nonres[insp_cols].sort_values(f'{insp}_spare', ascending=False).head()

''' 
----------------------------------
Get the non-residential growth summary by SA2
----------------------------------
'''
# Groupby and sum
sa2_nonres_alloc = pb_alloc_nonres.reset_index(drop=True).groupby('sa2_name').agg('sum')

# Inspect
sa2_nonres_alloc

# *Outputs*

# Output SA2 metrics to csv
sa2_nonres_alloc.to_csv(os.path.join(outputs_path, 'OUT_Redland_SA2_NonResGrowthAllocation.csv'))

# Merge with spatial
pb_alloc_nonres_out = pd.merge(pb_curdev_lyr.loc[:, ('pbno', 'SHAPE')], pb_alloc_nonres.reset_index(), on='pbno', how='left')

# Output into geodatabase
pb_alloc_nonres_out.spatial.to_featureclass(os.path.join(alloc_gdb_path, 'OUT_FC_PropertyBase_NonResGrowthAllocation'), overwrite=True)
