==> notebooks/P6_allocation_reporting_features.ipynb <==
# Establish connections to the input db and variables spreadsheets

# Import modules
import os
import arcpy
import pandas as pd
import numpy as np
from arcgis.features import GeoAccessor, GeoSeriesAccessor
from dotenv import load_dotenv

# Lot the .env variables
load_dotenv()

# Set the data paths
outputs_path = os.getenv('J111_OUTPUTS')
spatial_path = os.getenv('J111_SPATIAL')
env_path = os.getenv('J111_ENV')

# Get SA2 report outputs
res_csv = os.path.join(outputs_path, 'OUT_Redland_SA2_ResGrowthAllocation.csv')
nonres_csv = os.path.join(outputs_path, 'OUT_Redland_SA2_NonResGrowthAllocation.csv')

# Get full path of Inputs GDB
input_gdb_path = os.path.join(spatial_path, 'redland_gam_current_dev.gdb')

# Set arcpy workspace
arcpy.env.workspace = os.path.join(env_path, 'J111_redland_gam.gdb')

# Import the SA2 boundaries feature class
areas = pd.DataFrame.spatial.from_featureclass(os.path.join(input_gdb_path, 'INPUT_FC_SA2_Redland'))

# Keep only the necessary fields
areas = areas.loc[:, ('sa2_name_2016', 'SHAPE')].rename({'sa2_name_2016' : 'sa2_name'}, axis=1).set_index('sa2_name')

# Inspect dataframe
areas.head()

# Join residential data
res_data = pd.read_csv(res_csv).set_index('sa2_name')
res_areas = areas.join(res_data).reset_index()
res_areas.head()

# Join non-residential data
uses = [
    'commercial',
    'community',
    'education',
    'health',
    'industrial',
    'other',
    'retail',
]
horizons = ['2021', '2026', '2031', '2036', '2041', '2046', '2051', 'spare', 'cap']

nonres_data = pd.read_csv(nonres_csv).set_index('sa2_name')

# Get the largest contributing industry by horizon
new_cols = pd.DataFrame(np.empty([len(nonres_data), len(horizons)], dtype=str), columns=[f'largest_{h}' for h in horizons])
nonres_data = pd.concat([nonres_data, new_cols], axis=1)

for i, r in nonres_data.iterrows():

    for h in horizons:

        v = 0
        uv = None

        for u in uses:

            if h == 'cap':

                tv = r[f'{u}_2051'] + r[f'{u}_spare']

            else:

                tv = r[f'{u}_{h}']

            if tv > v:

                v = tv
                uv = u

        nonres_data.at[i, f'largest_{h}'] = uv

# Join to spatial
nonres_areas = areas.join(nonres_data).reset_index()
nonres_areas.head()

# Output
out_gdb = os.path.join(spatial_path, 'redland_gam_alloc.gdb')
res_areas.spatial.to_featureclass(os.path.join(out_gdb, 'OUT_FC_SA2_ResGrowthAllocation'), overwrite=True)
nonres_areas.spatial.to_featureclass(os.path.join(out_gdb, 'OUT_FC_SA2_NonResGrowthAllocation'), overwrite=True)
