==> notebooks/P8_idm_run_water_conversions.ipynb <==
# Establish connections to the input db and variables spreadsheets

# Import modules
import os, re, math
import arcpy
import pandas as pd
import numpy as np
from arcgis.features import GeoAccessor, GeoSeriesAccessor
from dotenv import load_dotenv

# Lot the .env variables
load_dotenv()

# Set the data paths
inputs_path = os.getenv('J111_INPUTS')
data_path = os.getenv('J122_DATA')
spatial_path = os.getenv('J111_SPATIAL')
env_path = os.getenv('J111_ENV')

# Get full path of Input GDBs
in_gdb_path = os.path.join(spatial_path, 'redland_gam_alloc.gdb')
out_gdb_path = os.path.join(spatial_path, 'redland_gam_idm.gdb')

# Set arcpy workspace
arcpy.env.workspace = os.path.join(env_path, 'J111_redland_gam.gdb')

# Inspect
in_gdb_path

''' 
----------------------------------
Import and prepare the input datasets
----------------------------------
'''

# Import the residential and non-residential development projections
res_dp = pd.DataFrame.spatial.from_featureclass(os.path.join(in_gdb_path, 'OUT_FC_PropertyBase_ResGrowthAllocation')).drop(['OBJECTID'], axis=1)
nonres_dp = pd.DataFrame.spatial.from_featureclass(os.path.join(in_gdb_path, 'OUT_FC_PropertyBase_NonResGrowthAllocation')).drop(['OBJECTID', 'SHAPE', 'sa2_name'], axis=1)

dp = pd.merge(left=res_dp, right=nonres_dp, on='pbno')

# Import current water consumption feature class
wc = pd.DataFrame.spatial.from_featureclass(os.path.join(out_gdb_path, 'OUT_FC_PropertyBase_WaterConsumption')).drop(['OBJECTID', 'SHAPE', 'landno', 'propno', 'segpar', 'prop_use'], axis=1)

# Inspect
dp.head()

''' 
----------------------------------
Get the contribution of existing non projected development as an ep
----------------------------------
'''
wc_nodev = wc.loc[:, ('pbno', 'nodev_wc')]

wc_nodev['ep_2021'] = ((wc_nodev['nodev_wc'] * 1000) / 365) / 230

wc_nodev = wc_nodev.drop(['nodev_wc'], axis=1).set_index('pbno').iloc[:,0]

wc_nodev

ep_rates = {
    "att_dwl" : 1.7,
    "det_dwl" : 2.7,
    "commerical" : 0.004,
    "community" : 0.006,
    "education" : 0.005,
    "health" : 0.012,
    "industrial" : 0.010,
    "other" : 0.006,
    "retail" : 0.005
}

ep_years = {
    '2021' : 'ep_2021', 
    '2026' : 'ep_2026', 
    '2031' : 'ep_2031' , 
    '2036' : 'ep_2036', 
    '2041' : 'ep_2041', 
    '2046' : 'ep_2046', 
    '2051' : 'ep_2051',
    'spare' : 'ep_ult'
}

dpep = dp.copy()

for suffix, name in ep_years.items():

    dpep[f'{name}'] = 0.0

for index, row in dpep.iterrows():

    for suffix, col_name in ep_years.items():

        ep = 0.0
        ep_nodev = wc_nodev[dpep.iloc[index, :]['pbno']]

        for col, val in row.iteritems():

            if col.endswith(suffix):

                for dev, rate in ep_rates.items():

                    if col.startswith(dev):

                        ep += round(val * rate, 2)

        if suffix == 'spare':

            ep += dpep.iloc[index, :]['ep_2051']

        ep += ep_nodev

        dpep.at[index, col_name] = ep

dpep = dpep.loc[:, ('pbno', 'sa2_name', *[name for suffix, name in ep_years.items()])]

dpep.head()

# Import the water zone details per property 

wsz = pd.DataFrame.spatial.from_table(os.path.join(out_gdb_path, 'INPUT_TB_Propbase_WaterSupplyZones')).drop(['OBJECTID'], axis=1)

wsz = wsz.loc[:, ('pbno', 'zone', 'prop_area', 'part_area')]

wsz = wsz.groupby(['pbno', 'zone', 'prop_area']).sum().reset_index()

wsz['portion'] = wsz.apply(lambda row: 1.0 if row['part_area'] == 0.0 else round(row['part_area'] / row['prop_area'], 2), axis=1)

wsz = wsz.loc[:, ('pbno', 'zone', 'portion')].drop_duplicates(subset=['pbno', 'zone', 'portion']).reset_index(drop=True)

wsz.head()

# Join ep demand to water supply zone properties

wsz_ep = pd.merge(left=wsz, right=dpep, on='pbno')

for suffix, col_name in ep_years.items():

    wsz_ep[f'{col_name}'] = round(wsz_ep[f'{col_name}'] * wsz_ep['portion'], 2)

wsz_ep = wsz_ep.loc[:, ('zone', *[name for suffix, name in ep_years.items()])].groupby('zone').sum()

wsz_ep.head()

# Import the sewer zone details per property

sz = pd.DataFrame.spatial.from_table(os.path.join(out_gdb_path, 'INPUT_TB_Propbase_SewerZones')).drop(['OBJECTID'], axis=1)

sz = sz.loc[:, ('pbno', 'zone', 'prop_area', 'part_area')]

sz = sz.groupby(['pbno', 'zone', 'prop_area']).sum().reset_index()

sz['portion'] = sz.apply(lambda row: 1.0 if row['part_area'] == 0.0 else round(row['part_area'] / row['prop_area'], 2), axis=1)

sz = sz.loc[:, ('pbno', 'zone', 'portion')].drop_duplicates(subset=['pbno', 'zone', 'portion']).set_index('pbno')

sz.head()

# Join ep demand to sewer supply zone properties

sz_ep = pd.merge(left=sz, right=dpep, on='pbno')

for suffix, col_name in ep_years.items():

    sz_ep[f'{col_name}'] = round(sz_ep[f'{col_name}'] * sz_ep['portion'], 2)

sz_ep = sz_ep.loc[:, ('zone', *[name for suffix, name in ep_years.items()])].groupby('zone').sum()

sz_ep.head()

# Get the ep demand per sa2

abs_ep = dpep.drop(['pbno'], axis=1).groupby('sa2_name').sum()

abs_ep

''' 
----------------------------------
Output EP summary tables
----------------------------------
'''

outputs_path = os.path.join(data_path, 'idm_outputs')

# Output metrics to csv
wsz_ep.to_csv(os.path.join(outputs_path, 'OUT_RedlandWater_IDM_EPDemandByWaterZone.csv'))
sz_ep.to_csv(os.path.join(outputs_path, 'OUT_RedlandWater_IDM_EPDemandBySewerZone.csv'))
abs_ep.to_csv(os.path.join(outputs_path, 'OUT_RedlandWater_IDM_EPDemandBySA2.csv'))

''' 
----------------------------------
Output EP Projections per Property
----------------------------------
'''

# Construt output featureclass
out_fc = pd.merge(left=res_dp.loc[:, ('pbno', 'SHAPE')], right=dpep, on='pbno')

# Output into geodatabase
out_fc.spatial.to_featureclass(os.path.join(out_gdb_path, 'OUT_FC_PropertyBase_EPDemand'), overwrite=True)
